+++ src/Controller/Api/Stations/MountsController.php
@@ -6,9 +6,15 @@ namespace App\Controller\Api\Stations;
 
 use App\Entity;
 use App\Exception\StationUnsupportedException;
+use App\Http\Response;
 use App\Http\Router;
 use App\Http\ServerRequest;
+use App\Service\Flow\UploadedFile;
+use Doctrine\ORM\EntityManagerInterface;
 use OpenApi\Annotations as OA;
+use Psr\Http\Message\ResponseInterface;
+use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
  * @extends AbstractStationApiCrudController<Entity\StationMount>
@@ -17,6 +23,15 @@ class MountsController extends AbstractStationApiCrudController
 {
     protected string $entityClass = Entity\StationMount::class;
     protected string $resourceRouteName = 'api:stations:mount';
+
+    public function __construct(
+        EntityManagerInterface $em,
+        Serializer $serializer,
+        ValidatorInterface $validator,
+        protected Entity\Repository\StationMountRepository $mountRepo
+    ) {
+        parent::__construct($em, $serializer, $validator);
+    }
 
     /**
      * @OA\Get(path="/station/{station_id}/mounts",
@@ -104,11 +119,18 @@ class MountsController extends AbstractStationApiCrudController
 
     protected function viewRecord(object $record, ServerRequest $request): mixed
     {
+        /** @var Entity\StationMount $record */
         $return = parent::viewRecord($record, $request);
 
         $station = $request->getStation();
         $frontend = $request->getStationFrontend();
         $router = $request->getRouter();
+
+        $return['links']['intro'] = (string)$router->fromHere(
+            route_name:   'api:stations:mounts:intro',
+            route_params: ['id' => $record->getId()],
+            absolute:     true
+        );
 
         $return['links']['listen'] = (string)Router::resolveUri(
             $router->getBaseUrl(),
@@ -117,6 +139,44 @@ class MountsController extends AbstractStationApiCrudController
         );
 
         return $return;
+    }
+
+    public function createAction(
+        ServerRequest $request,
+        Response $response
+    ): ResponseInterface {
+        $station = $request->getStation();
+
+        $parsedBody = (array)$request->getParsedBody();
+        $record = $this->editRecord(
+            $parsedBody,
+            new Entity\StationMount($station)
+        );
+
+        if (!empty($parsedBody['intro_file'])) {
+            $intro = UploadedFile::fromArray($parsedBody['intro_file'], $station->getRadioTempDir());
+            $this->mountRepo->setIntro($record, $intro);
+        }
+
+        return $response->withJson($this->viewRecord($record, $request));
+    }
+
+    public function deleteAction(
+        ServerRequest $request,
+        Response $response,
+        int $station_id,
+        int $id
+    ): ResponseInterface {
+        $record = $this->getRecord($this->getStation($request), $id);
+
+        if (null === $record) {
+            return $response->withStatus(404)
+                ->withJson(Entity\Api\Error::notFound());
+        }
+
+        $this->mountRepo->destroy($record);
+
+        return $response->withJson(new Entity\Api\Status(true, __('Record deleted successfully.')));
     }
 
     /**